requirements.txt
apache-airflow-providers-google
astronomer-cosmos[dbt-bigquery]
protobuf